requirements.txt
openai
python-dotenv
ollama
